fix(api): update venue data in seat availability calculation
The changes here mainly concern the AvailableSeatsCalculator service in the stt app. Specifically, the code for retrieving venue information has been modified. Now, instead of returning the home_venue object, it returns the name of the home_venue. This is done so that there will be no breaking changes for the frontend.