style(api): reorder import statements in serializers
The import statements for the 'get_user_model', 'authenticate', 'password_validation', 'localtime', and 'ContentType' from 'django.contrib.auth' and 'django.utils.timezone' have been moved to the top of the import list in the serializers.py files for the 'users', 'stt', 'email_domains', and 'in_app_notifications' apps respectively. This change is in adherence to PEP8 guidelines that recommend standard library imports should be on top of other imports.